cree nuevos modelos y separece el auth_controller.py para mejor visualisacion y mejor entendimiento
