+++ src/adapters/primary/http/routes/auth_routes.py
@@ -1,6 +1,6 @@
 from flask import Blueprint
-from src.adapters.primary.http.controllers.auth_controller import auth_blueprint
+from src.adapters.primary.http.controllers.user_module import user_module
 
 # Registrar el blueprint de autenticación
 auth_routes = Blueprint('auth_routes', __name__)
-auth_routes.register_blueprint(auth_blueprint, url_prefix='/auth') 
+auth_routes.register_blueprint(user_module, url_prefix='/auth')
